requirements.txt
amqp==1.4.9
ansible==1.9.6
anyjson==0.3.3
apache-libcloud==0.17.0
apipkg==1.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
boto==2.38.0
cairocffi==0.8.0
CairoSVG==1.0.22
celery==3.1.19
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
CommonMark==0.7.4
coverage==3.7.1
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.4.0
Django==1.8.18
django-allauth==0.19.1
django-braces==1.11.0
django-cors-headers==1.0.0
django-debug-toolbar==1.3.2
django-environ==0.3.1
django-filter==0.15.3
django-oauth-toolkit==0.8.2
django-redis==4.0.0
django-rest-auth==0.4.0
django-rest-swagger==0.2.9
django-reversion==2.0.5
django-storages==1.1.8
djangorestframework==3.1.3
djangorestframework-csv==1.3.4
djangorestframework-gis==0.8.2
djangorestframework-xml==1.0.1
docutils==0.14
drf-extensions==0.2.8
ecdsa==0.13
execnet==1.5.0
Fabric==1.10.5
filechunkio==1.6
flake8==2.3.0
future==0.16.0
gunicorn==19.3.0
html5lib==0.999999999
idna==2.6
Jinja2==2.10
kombu==3.0.37
Markdown==2.5.2
MarkupSafe==1.0
mccabe==0.6.1
-e git+https://github.com/uonafya/mfl_api.git@3241c61507fecbc83fd8cc28d3d72be209c92f91#egg=mfl-mapping_models
mock==1.0.1
model-mommy==1.2.6
oauthlib==2.0.6
paramiko==1.18.4
pdfrw==0.4
pep8==1.7.1
pkg-resources==0.0.0
psycopg2==2.5.5
py==1.5.2
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.6.0
Pyphen==0.9.4
pytest==2.7.3
pytest-django==2.8.0
pytest-xdist==1.11
python-coveralls==2.5.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.3
PyYAML==3.12
recommonmark==0.1.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
sh==1.12.14
Shapely==1.5.17.post1
six==1.9.0
sqlparse==0.1.19
tinycss2==0.6.1
tox==1.9.2
urllib3==1.22
virtualenv==12.0.7
WeasyPrint==0.41
webencodings==0.5.1
Werkzeug==0.10.4
wheel==0.30.0
XlsxWriter==0.7.9
